fixing the problem with pending notifications disappearing when starting the application (not from memory)
https://github.com/dpa99c/cordova-plugin-firebasex/issues/272

https://github.com/dpa99c/cordova-plugin-firebasex/issues/674

+++ src/android/FirebasePlugin.java
@@ -20,6 +20,9 @@ import androidx.core.app.ActivityCompat;
 import androidx.core.app.NotificationCompat;
 import androidx.core.app.NotificationManagerCompat;
 
+import android.service.notification.StatusBarNotification;
+import android.app.Notification;
+
 import android.util.Base64;
 import android.util.Log;
 
@@ -2717,7 +2720,7 @@ public class FirebasePlugin extends CordovaPlugin {
             String id = options.getString("id");
             Log.i(TAG, "Creating channel id=" + id);
 
-            if (channelExists(id)) {
+            if( channelExists(id) && !channelHasNotifications(id) ) {
                 deleteChannel(id);
             }
 
@@ -2815,6 +2818,25 @@ public class FirebasePlugin extends CordovaPlugin {
             nm.createNotificationChannel(channel);
         }
         return channel;
+    }
+    
+    protected static boolean channelHasNotifications(String channelId){
+        if(Build.VERSION.SDK_INT < Build.VERSION_CODES.O){
+            return false;
+        }
+
+        NotificationManager nm = (NotificationManager) applicationContext.getSystemService(Context.NOTIFICATION_SERVICE);
+        StatusBarNotification[] notifications = nm.getActiveNotifications();
+
+        for(int i = 0; i < notifications.length; i++){
+            Notification notification = notifications[i].getNotification();
+
+            if(notification.getChannelId().equals(channelId)){
+                return true;
+            }
+        }
+
+        return false;
     }
 
     protected static void createDefaultChannel() throws JSONException {
